Routing and DPP modifications
The numberRequests variable's type was changed. 
ProtectDisjointYEN was changed to consider as disjoint routes as possible instead Kd routes.
The ProtectYEN routing algorithm was created.
Some adjusts in DPP was performed.

+++ include/SimulationType/SimulationType.h
@@ -179,7 +179,7 @@ public:
      * This variable is updated when a new call/event is create,
      * in function GenerateCall of CallGenerator class.
      */
-    double numberRequests;
+    unsigned long long int numberRequests;
 private:
     /**
      * @brief Define the type of simulation based on the option chosen.
